+++ internal/cloud/backup.go
@@ -1,6 +1,10 @@
 package cloud
 
-import "time"
+import (
+	"fmt"
+	"strings"
+	"time"
+)
 
 // Backup store all the necessary information of an already uploaded archive.
 type Backup struct {
@@ -19,4 +23,40 @@ type Backup struct {
 
 	// Size backup archive size.
 	Size int64
+
+	// Location defines where the backup was stored.
+	Location Location
+}
+
+const (
+	// LocationAWS indicates that the backup was stored in Amazon AWS cloud.
+	LocationAWS Location = "aws"
+
+	// LocationGCS indicates that the backup was stored in Google Cloud Storage.
+	LocationGCS Location = "gcs"
+)
+
+// Location contains the cloud that is current storing the backup data.
+type Location string
+
+// ParseLocation converts a text to a Location type.
+func ParseLocation(value string) (Location, error) {
+	value = strings.ToLower(value)
+	value = strings.TrimSpace(value)
+
+	switch value {
+	case string(LocationAWS):
+		return LocationAWS, nil
+	case string(LocationGCS):
+		return LocationGCS, nil
+	}
+
+	// not return a library error here because this is used by the library itself
+	// to build backups from storage
+	return Location(""), fmt.Errorf("unknown location “%s”", value)
+}
+
+// Defined returns true if the location has a valid value.
+func (l Location) Defined() bool {
+	return l == LocationAWS || l == LocationGCS
 }