+++ internal/storage/audit_file_test.go
@@ -54,9 +54,10 @@ func TestAuditFile_Save(t *testing.T) {
 					Checksum:  "ca34f069795292e834af7ea8766e9e68fdddf3f46c7ce92ab94fc2174910adb7",
 					VaultName: "test",
 					Size:      120,
+					Location:  cloud.LocationAWS,
 				},
 			},
-			expected: fmt.Sprintf("%s test 123456 ca34f069795292e834af7ea8766e9e68fdddf3f46c7ce92ab94fc2174910adb7 120\n", now.Format(time.RFC3339)),
+			expected: fmt.Sprintf("%s test 123456 ca34f069795292e834af7ea8766e9e68fdddf3f46c7ce92ab94fc2174910adb7 120 aws\n", now.Format(time.RFC3339)),
 		},
 		{
 			description: "it should detect when the filename refers to a directory",
@@ -79,6 +80,7 @@ func TestAuditFile_Save(t *testing.T) {
 					CreatedAt: now,
 					Checksum:  "ca34f069795292e834af7ea8766e9e68fdddf3f46c7ce92ab94fc2174910adb7",
 					VaultName: "test",
+					Location:  cloud.LocationAWS,
 				},
 			},
 			expectedError: &storage.Error{
@@ -138,8 +140,8 @@ func TestAuditFile_List(t *testing.T) {
 				}
 				defer f.Close()
 
-				f.WriteString(fmt.Sprintf("%s test 123456 ca34f069795292e834af7ea8766e9e68fdddf3f46c7ce92ab94fc2174910adb7 120\n", now.Format(time.RFC3339)))
-				f.WriteString(fmt.Sprintf("%s test 654321 ca34f069795292e834af7ea8766e9e68fdddf3f46c7ce92ab94fc2174910adb7 120\n", now.Add(time.Second).Format(time.RFC3339)))
+				f.WriteString(fmt.Sprintf("%s test 123456 ca34f069795292e834af7ea8766e9e68fdddf3f46c7ce92ab94fc2174910adb7 120 aws\n", now.Format(time.RFC3339)))
+				f.WriteString(fmt.Sprintf("%s test 654321 ca34f069795292e834af7ea8766e9e68fdddf3f46c7ce92ab94fc2174910adb7 120 aws\n", now.Add(time.Second).Format(time.RFC3339)))
 				return f.Name()
 			}(),
 			expected: storage.Backups{
@@ -156,6 +158,7 @@ func TestAuditFile_List(t *testing.T) {
 						Checksum:  "ca34f069795292e834af7ea8766e9e68fdddf3f46c7ce92ab94fc2174910adb7",
 						VaultName: "test",
 						Size:      120,
+						Location:  cloud.LocationAWS,
 					},
 				},
 				{
@@ -171,6 +174,7 @@ func TestAuditFile_List(t *testing.T) {
 						Checksum:  "ca34f069795292e834af7ea8766e9e68fdddf3f46c7ce92ab94fc2174910adb7",
 						VaultName: "test",
 						Size:      120,
+						Location:  cloud.LocationAWS,
 					},
 				},
 			},
@@ -206,12 +210,13 @@ func TestAuditFile_List(t *testing.T) {
 						}(),
 						Checksum:  "ca34f069795292e834af7ea8766e9e68fdddf3f46c7ce92ab94fc2174910adb7",
 						VaultName: "test",
+						Location:  cloud.LocationAWS,
 					},
 				},
 			},
 		},
 		{
-			description: "it should list all backups information correctly with backward compatibility",
+			description: "it should list all backups information correctly with backward compatibility (no size)",
 			logger: mockLogger{
 				mockDebug:  func(args ...interface{}) {},
 				mockDebugf: func(format string, args ...interface{}) {},
@@ -241,6 +246,44 @@ func TestAuditFile_List(t *testing.T) {
 						}(),
 						Checksum:  "ca34f069795292e834af7ea8766e9e68fdddf3f46c7ce92ab94fc2174910adb7",
 						VaultName: "test",
+						Location:  cloud.LocationAWS,
+					},
+				},
+			},
+		},
+		{
+			description: "it should list all backups information correctly with backward compatibility (no location)",
+			logger: mockLogger{
+				mockDebug:  func(args ...interface{}) {},
+				mockDebugf: func(format string, args ...interface{}) {},
+				mockInfo:   func(args ...interface{}) {},
+				mockInfof:  func(format string, args ...interface{}) {},
+			},
+			filename: func() string {
+				f, err := ioutil.TempFile("", "toglacier-test")
+				if err != nil {
+					t.Fatalf("error creating a temporary file. details: %s", err)
+				}
+				defer f.Close()
+
+				f.WriteString(fmt.Sprintf("%s test 123456 ca34f069795292e834af7ea8766e9e68fdddf3f46c7ce92ab94fc2174910adb7 120\n", now.Format(time.RFC3339)))
+				return f.Name()
+			}(),
+			expected: storage.Backups{
+				{
+					Backup: cloud.Backup{
+						ID: "123456",
+						CreatedAt: func() time.Time {
+							c, err := time.Parse(time.RFC3339, now.Format(time.RFC3339))
+							if err != nil {
+								t.Fatalf("error parsing current time. details: %s", err)
+							}
+							return c
+						}(),
+						Checksum:  "ca34f069795292e834af7ea8766e9e68fdddf3f46c7ce92ab94fc2174910adb7",
+						VaultName: "test",
+						Size:      120,
+						Location:  cloud.LocationAWS,
 					},
 				},
 			},
@@ -274,7 +317,7 @@ func TestAuditFile_List(t *testing.T) {
 					}
 					defer f.Close()
 
-					f.WriteString(fmt.Sprintf("%s test 123456 ca34f069795292e834af7ea8766e9e68fdddf3f46c7ce92ab94fc2174910adb7 120\n", now.Format(time.RFC3339)))
+					f.WriteString(fmt.Sprintf("%s test 123456 ca34f069795292e834af7ea8766e9e68fdddf3f46c7ce92ab94fc2174910adb7 120 aws\n", now.Format(time.RFC3339)))
 				}
 
 				return n
@@ -349,7 +392,7 @@ func TestAuditFile_List(t *testing.T) {
 				}
 				defer f.Close()
 
-				f.WriteString("XXXX test 123456 ca34f069795292e834af7ea8766e9e68fdddf3f46c7ce92ab94fc2174910adb7 120\n")
+				f.WriteString("XXXX test 123456 ca34f069795292e834af7ea8766e9e68fdddf3f46c7ce92ab94fc2174910adb7 120 aws\n")
 				return f.Name()
 			}(),
 			expectedError: &storage.Error{
@@ -378,7 +421,7 @@ func TestAuditFile_List(t *testing.T) {
 				}
 				defer f.Close()
 
-				f.WriteString(fmt.Sprintf("%s test 123456 ca34f069795292e834af7ea8766e9e68fdddf3f46c7ce92ab94fc2174910adb7 XXXX\n", now.Format(time.RFC3339)))
+				f.WriteString(fmt.Sprintf("%s test 123456 ca34f069795292e834af7ea8766e9e68fdddf3f46c7ce92ab94fc2174910adb7 XXXX aws\n", now.Format(time.RFC3339)))
 				return f.Name()
 			}(),
 			expectedError: &storage.Error{
@@ -388,6 +431,29 @@ func TestAuditFile_List(t *testing.T) {
 					Num:  "XXXX",
 					Err:  errors.New("invalid syntax"),
 				},
+			},
+		},
+		{
+			description: "it should detect when the audit file contains an invalid location",
+			logger: mockLogger{
+				mockDebug:  func(args ...interface{}) {},
+				mockDebugf: func(format string, args ...interface{}) {},
+				mockInfo:   func(args ...interface{}) {},
+				mockInfof:  func(format string, args ...interface{}) {},
+			},
+			filename: func() string {
+				f, err := ioutil.TempFile("", "toglacier-test")
+				if err != nil {
+					t.Fatalf("error creating a temporary file. details: %s", err)
+				}
+				defer f.Close()
+
+				f.WriteString(fmt.Sprintf("%s test 123456 ca34f069795292e834af7ea8766e9e68fdddf3f46c7ce92ab94fc2174910adb7 120 XXXX\n", now.Format(time.RFC3339)))
+				return f.Name()
+			}(),
+			expectedError: &storage.Error{
+				Code: storage.ErrorCodeLocation,
+				Err:  errors.New("unknown location “xxxx”"),
 			},
 		},
 	}
@@ -434,15 +500,15 @@ func TestAuditFile_Remove(t *testing.T) {
 				}
 				defer f.Close()
 
-				f.WriteString(fmt.Sprintf("%s test 123456 ca34f069795292e834af7ea8766e9e68fdddf3f46c7ce92ab94fc2174910adb7 100\n", now.Format(time.RFC3339)))
-				f.WriteString(fmt.Sprintf("%s test 123457 913b87897ffb6dca07e9f17e280aa8ecb9886dffeda8a15efeafec11dec0d108 200\n", now.Add(time.Second).Format(time.RFC3339)))
+				f.WriteString(fmt.Sprintf("%s test 123456 ca34f069795292e834af7ea8766e9e68fdddf3f46c7ce92ab94fc2174910adb7 100 aws\n", now.Format(time.RFC3339)))
+				f.WriteString(fmt.Sprintf("%s test 123457 913b87897ffb6dca07e9f17e280aa8ecb9886dffeda8a15efeafec11dec0d108 200 aws\n", now.Add(time.Second).Format(time.RFC3339)))
 				return f.Name()
 			}(),
 			id:       "123457",
-			expected: fmt.Sprintf("%s test 123456 ca34f069795292e834af7ea8766e9e68fdddf3f46c7ce92ab94fc2174910adb7 100\n", now.Format(time.RFC3339)),
+			expected: fmt.Sprintf("%s test 123456 ca34f069795292e834af7ea8766e9e68fdddf3f46c7ce92ab94fc2174910adb7 100 aws\n", now.Format(time.RFC3339)),
 		},
 		{
-			description: "it should remove a backup information correctly with backward compatibility",
+			description: "it should remove a backup information correctly with backward compatibility (no size)",
 			logger: mockLogger{
 				mockDebug:  func(args ...interface{}) {},
 				mockDebugf: func(format string, args ...interface{}) {},
@@ -461,7 +527,29 @@ func TestAuditFile_Remove(t *testing.T) {
 				return f.Name()
 			}(),
 			id:       "123457",
-			expected: fmt.Sprintf("%s test 123456 ca34f069795292e834af7ea8766e9e68fdddf3f46c7ce92ab94fc2174910adb7 0\n", now.Format(time.RFC3339)),
+			expected: fmt.Sprintf("%s test 123456 ca34f069795292e834af7ea8766e9e68fdddf3f46c7ce92ab94fc2174910adb7 0 aws\n", now.Format(time.RFC3339)),
+		},
+		{
+			description: "it should remove a backup information correctly with backward compatibility (no location)",
+			logger: mockLogger{
+				mockDebug:  func(args ...interface{}) {},
+				mockDebugf: func(format string, args ...interface{}) {},
+				mockInfo:   func(args ...interface{}) {},
+				mockInfof:  func(format string, args ...interface{}) {},
+			},
+			filename: func() string {
+				f, err := ioutil.TempFile("", "toglacier-test")
+				if err != nil {
+					t.Fatalf("error creating a temporary file. details: %s", err)
+				}
+				defer f.Close()
+
+				f.WriteString(fmt.Sprintf("%s test 123456 ca34f069795292e834af7ea8766e9e68fdddf3f46c7ce92ab94fc2174910adb7 120\n", now.Format(time.RFC3339)))
+				f.WriteString(fmt.Sprintf("%s test 123457 913b87897ffb6dca07e9f17e280aa8ecb9886dffeda8a15efeafec11dec0d108 120\n", now.Add(time.Second).Format(time.RFC3339)))
+				return f.Name()
+			}(),
+			id:       "123457",
+			expected: fmt.Sprintf("%s test 123456 ca34f069795292e834af7ea8766e9e68fdddf3f46c7ce92ab94fc2174910adb7 120 aws\n", now.Format(time.RFC3339)),
 		},
 		{
 			description: "it should detect when the audit file has no read permission",
@@ -480,7 +568,7 @@ func TestAuditFile_Remove(t *testing.T) {
 					}
 					defer f.Close()
 
-					f.WriteString(fmt.Sprintf("%s test 123456 ca34f069795292e834af7ea8766e9e68fdddf3f46c7ce92ab94fc2174910adb7\n", now.Format(time.RFC3339)))
+					f.WriteString(fmt.Sprintf("%s test 123456 ca34f069795292e834af7ea8766e9e68fdddf3f46c7ce92ab94fc2174910adb7 aws\n", now.Format(time.RFC3339)))
 				}
 
 				return n
@@ -508,7 +596,7 @@ func TestAuditFile_Remove(t *testing.T) {
 			}
 
 			if !reflect.DeepEqual(scenario.expected, string(auditFileContent)) {
-				t.Errorf("audit file don't match. expected “%s” and got “%s”", scenario.expectedError, string(auditFileContent))
+				t.Errorf("audit file don't match. expected “%v” and got “%v”", scenario.expected, string(auditFileContent))
 			}
 
 			if !storage.ErrorEqual(scenario.expectedError, err) {